organize llm adapters
