+++ python/api_adapters/gemini_service_adapter.py
