+++ python/api_adapters/openai_service_adapter.py
